dap: Fix an off-by-one and move the function over to commands/dap
